don't require master password in opts
opts.masterPassword is still required for initialization, but now
once Shroud is initialized, the methods that required the master pasword
expect it to be passed explicitly as arguments.

This allows for a number of things to happen without authentication:
adding a new secret, removing a secret (maybe this isn't a good idea?),
and retrieving raw key data from the vault and data files.